requirements.txt
Django==2.0.2
django-micro==1.6.1
djangorestframework==3.7.7
